editing views 11

+++ database/migrations/2020_07_19_161922_create_view_users_posts_up_ul_view.php
@@ -37,7 +37,7 @@ class CreateViewUsersPostsUpUlView extends Migration
             users
 
         LEFT JOIN
-            user_lands ON users.id = user_lands.user_id
+            user_lands ON users.id::integer = user_lands.user_id::integer
 
         LEFT JOIN
             posts ON users.id::integer = posts.user_id::integer
